Use resampling to force continuous queries to re-run
This is a more automated way of solving the backfilling issue, although
it won't help if the backfill is from a long time ago -- manual intervention
would be a better solution in that case.

Getting influx to handle this might be better for memory consumption than our
enforced backfilling, but we'll wait and see on that one.