Add lossrate and throughput rate metrics to data tables
The relative matrix for loss and throughput requires the stddev
of these rates. These can't be derived from our existing CQ
results, so we need to explicitly add them to the CQ.

Unfortunately influx is a bit picky about you trying to do math
in a CQ, so we're playing it safe and adding an extra column to
the data tables to precompute and store the result of the math.

We can then get influx to put the stddev of the resulting rate
into the matrix CQ tables.