Catch errors when dropping continuous queries
Influx now (correctly) throws an error if you try to drop
a CQ that doesn't exist. Unfortunately, we were relying on influx
ignoring that problem when configuring a DB for the first time.

Now we detect and silently ignore the error (because we only care
that the CQ isn't there, not whether it was there in the first
place).